requirements.txt
aiohttp==3.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.10.4
astroid==3.0.3
async-timeout==5.0.1
attrs==25.4.0
babel==2.17.0
bcrypt==5.0.0
beautifulsoup4==4.14.2
black==23.11.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coverage==7.11.3
cryptography==46.0.3
Deprecated==1.3.1
dill==0.4.0
ecdsa==0.19.1
environs==14.5.0
et_xmlfile==2.0.0
fastapi==0.104.1
fastapi_cors==0.0.6
flake8==6.1.0
frozenlist==1.8.0
ghp-import==2.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==21.2.0
h11==0.16.0
hiredis==2.2.3
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.25.2
idna==3.11
iniconfig==2.3.0
isort==5.13.2
Jinja2==3.1.6
limits==5.6.0
lxml==4.9.3
Markdown==3.10
MarkupSafe==3.0.3
marshmallow==4.1.0
mccabe==0.7.0
mergedeep==1.3.4
mkdocs==1.5.3
mkdocs-material==9.5.2
mkdocs-material-extensions==1.3.1
multidict==6.7.0
mypy==1.7.1
mypy_extensions==1.1.0
numpy==1.26.2
openpyxl==3.1.2
packaging==25.0
paginate==0.5.7
pandas==2.3.3
passlib==1.7.4
pathspec==0.12.1
pdfminer.six==20221105
pdfplumber==0.10.3
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pycparser==2.23
pydantic==2.5.0
pydantic_core==2.14.1
pyflakes==3.1.0
Pygments==2.19.2
pylint==3.0.3
pymdown-extensions==10.17.1
pyparsing==3.2.5
pypdf==3.17.1
PyPDF2==3.0.1
pypdfium2==5.0.0
pytesseract==0.3.13
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.3
pyyaml_env_tag==1.1
redis==5.0.1
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
selectolax==0.3.17
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.8
starlette==0.27.0
tenacity==8.2.3
tomlkit==0.13.3
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.24.0
validators==0.22.0
watchdog==6.0.0
watchfiles==1.1.1
webencodings==0.5.1
websockets==15.0.1
wrapt==2.0.1
XlsxWriter==3.1.9
yarl==1.22.0
